Moves applicable Domain Access tests to Functional.
