+++ domain/src/Tests/DomainTestBase.php
@@ -99,5 +99,4 @@ abstract class DomainTestBase extends WebTestBase {
     }
   }
 
-
 }
